% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatriceCooccurence_function.R
\name{MatriceCooccurence}
\alias{MatriceCooccurence}
\title{Fonction pour le calcul de la matrice de cooccurence}
\usage{
MatriceCooccurence(mots, presencemot)
}
\arguments{
\item{mots, presencemot}{.}
}
\description{
This function allows you to calculate the cooccurrence matrix.
}
\examples{
MatriceCooccurence()
}
\keyword{cooccurrence}
\keyword{matrix}
