% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probrencontre_function.R
\name{Probrencontre}
\alias{Probrencontre}
\title{Calculate probability of appearance}
\usage{
Probrencontre(N, n, f)
}
\description{
This function allows you to calculate the probability ofappearance of aword.
}
\examples{
Probrencontre()
}
\keyword{appearance}
